Change semantc sof highlighting in nodelink
Previously, clicking on a node sent a highlight message listing the
id of the node that was clicked on only; a halo ws drawn for highlighted
nodes and their neiighbours.

Now, clicking on a node sends a message including the id of that node
and its neighbours, and a halo is only drawn for nodes in this list.

This enables NetworkNarratives to highlight a specific set of nodes,
without affecting their neighbours.